+++ src/selectors/disabled.js
@@ -1,0 +1,65 @@
+import { createSelector } from 'reselect';
+import { arrayToObject } from '../utils';
+import { getTagCount } from './tags';
+
+const getNodeIDs = state => state.node.ids;
+const getNodeDisabledNode = state => state.node.disabled;
+const getNodeTags = state => state.node.tags;
+const getNodeType = state => state.node.type;
+const getTagEnabled = state => state.tag.enabled;
+const getNodeTypeDisabled = state => state.nodeType.disabled;
+const getEdgeIDs = state => state.edge.ids;
+const getEdgeSources = state => state.edge.sources;
+const getEdgeTargets = state => state.edge.targets;
+
+/**
+ * Calculate whether nodes should be disabled based on their tags
+ */
+export const getNodeDisabledTag = createSelector(
+  [getNodeIDs, getTagEnabled, getTagCount, getNodeTags],
+  (nodeIDs, tagEnabled, tagCount, nodeTags) =>
+    arrayToObject(nodeIDs, nodeID => {
+      if (tagCount.enabled === 0) {
+        return false;
+      }
+      if (nodeTags[nodeID].length) {
+        // Hide task nodes that don't have at least one tag filter enabled
+        return !nodeTags[nodeID].some(tag => tagEnabled[tag]);
+      }
+      return true;
+    })
+);
+
+/**
+ * Set disabled status if the node is specifically hidden, and/or via a tag/view/type
+ */
+export const getNodeDisabled = createSelector(
+  [
+    getNodeIDs,
+    getNodeDisabledNode,
+    getNodeDisabledTag,
+    getNodeType,
+    getNodeTypeDisabled
+  ],
+  (nodeIDs, nodeDisabledNode, nodeDisabledTag, nodeType, typeDisabled) =>
+    arrayToObject(nodeIDs, id =>
+      Boolean(
+        nodeDisabledNode[id] ||
+          nodeDisabledTag[id] ||
+          typeDisabled[nodeType[id]]
+      )
+    )
+);
+
+/**
+ * Determine whether an edge should be disabled based on their source/target nodes
+ */
+export const getEdgeDisabled = createSelector(
+  [getEdgeIDs, getNodeDisabled, getEdgeSources, getEdgeTargets],
+  (edgeIDs, nodeDisabled, edgeSources, edgeTargets) =>
+    arrayToObject(edgeIDs, edgeID => {
+      const source = edgeSources[edgeID];
+      const target = edgeTargets[edgeID];
+      return Boolean(nodeDisabled[source] || nodeDisabled[target]);
+    })
+);
