+++ src/selectors/nodes.test.js
@@ -1,7 +1,5 @@
 import { mockState } from '../utils/state.mock';
 import {
-  getNodeDisabledTag,
-  getNodeDisabled,
   getNodeActive,
   getNodeData,
   getNodeTextWidth,
@@ -15,84 +13,12 @@ import {
   toggleNodeHovered,
   toggleNodesDisabled
 } from '../actions/nodes';
-import { toggleTagFilter } from '../actions/tags';
 import reducer from '../reducers';
 
 const getNodeIDs = state => state.node.ids;
 const getNodeName = state => state.node.name;
-const getNodeTags = state => state.node.tags;
-const getNodeType = state => state.node.type;
 
 describe('Selectors', () => {
-  describe('getNodeDisabledTag', () => {
-    it('returns an object', () => {
-      expect(getNodeDisabledTag(mockState.lorem)).toEqual(expect.any(Object));
-    });
-
-    it("returns an object whose keys match the current pipeline's nodes", () => {
-      expect(Object.keys(getNodeDisabledTag(mockState.lorem))).toEqual(
-        getNodeIDs(mockState.lorem)
-      );
-    });
-
-    it('returns an object whose values are all Booleans', () => {
-      expect(
-        Object.values(getNodeDisabledTag(mockState.lorem)).every(
-          value => typeof value === 'boolean'
-        )
-      ).toBe(true);
-    });
-
-    it('does not disable a node if all tags are disabled', () => {
-      const nodeDisabled = getNodeDisabledTag(mockState.lorem);
-      expect(Object.values(nodeDisabled)).toEqual(
-        Object.values(nodeDisabled).map(() => false)
-      );
-    });
-
-    it('disables a node only if all of its tags are disabled', () => {
-      const nodeTags = getNodeTags(mockState.animals);
-      // Get list of task nodes from the current pipeline
-      const taskNodes = getNodeIDs(mockState.animals).filter(
-        id => getNodeType(mockState.animals)[id] === 'task'
-      );
-      // Choose a node that has some tags (and which should be enabled)
-      const hasTags = id => Boolean(nodeTags[id].length);
-      const enabledNodeID = taskNodes.find(hasTags);
-      // Choose a node that has no tags (and which should be disabled)
-      const hasNoTags = id => !Boolean(nodeTags[id].length);
-      const disabledNodeID = taskNodes.find(hasNoTags);
-      // Update the state to enable one of the tags for that node
-      const enabledNodeTags = nodeTags[enabledNodeID];
-      const newMockState = reducer(
-        mockState.animals,
-        toggleTagFilter(enabledNodeTags[0], true)
-      );
-      expect(getNodeDisabledTag(newMockState)[enabledNodeID]).toEqual(false);
-      expect(getNodeDisabledTag(newMockState)[disabledNodeID]).toEqual(true);
-    });
-  });
-
-  describe('getNodeDisabled', () => {
-    it('returns an object', () => {
-      expect(getNodeDisabled(mockState.lorem)).toEqual(expect.any(Object));
-    });
-
-    it("returns an object whose keys match the current pipeline's nodes", () => {
-      expect(Object.keys(getNodeDisabled(mockState.lorem))).toEqual(
-        getNodeIDs(mockState.lorem)
-      );
-    });
-
-    it('returns an object whose values are all Booleans', () => {
-      expect(
-        Object.values(getNodeDisabled(mockState.lorem)).every(
-          value => typeof value === 'boolean'
-        )
-      ).toBe(true);
-    });
-  });
-
   describe('getNodeActive', () => {
     it('returns an object', () => {
       expect(getNodeActive(mockState.lorem)).toEqual(expect.any(Object));
